backend/rest-api/requirements/requirements.txt
falcon==2.0.0
gunicorn
uwsgi
waitress
SQLAlchemy
python-rapidjson
loguru
python-environ
mysql-connector-python